+++ frontend/src/components/ParkingStatus.tsx
@@ -1,0 +1,10 @@
+const ParkingStatus = () => {
+  return (
+    <div className="flex w-full flex-row justify-evenly p-5">
+      <h1 className="text-5xl"> ที่จอดรถว่าง 5 คัน </h1>
+      <h1 className="text-5xl"> รถขับวนอยู่ 2 คัน </h1>
+    </div>
+  );
+};
+
+export default ParkingStatus;
